Admin

+++ src/main/java/com/example/approval24/domain/InstDTO.java
@@ -9,7 +9,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class InstDTO {
 	
-	private int instId;
+	private long instId;
 	
 	private String instName;
